refactor: relocate SortableTableHead, drop Readonly<Props>, extract formatExpiry (#509, #582, #601)
- #509: SortableTableHead is consumed across feature boundaries (admin/dashboard
  + reports), so move it out of admin/dashboard/_lib into the shared
  apps/web/components/sortable-table-head.tsx. Update all 3 importers + the test.
- #582: drop the Readonly<Props> wrapper from the 3 internal-exam components that
  still used it (attempts-table, issue-code-form, void-code-dialog) for
  consistency with the siblings already on plain Props. (#582 said 5; only 3
  remained.)
- #601: formatExpiry was exported from issued-code-panel.tsx only so the test
  could import it (widening the component's public API). Move it to a co-located
  _utils/format-expiry.ts with its own test; the component now imports it.

No behavior changes.

Closes #509
Closes #582
Closes #601
