+++ apps/web/app/app/admin/dashboard/_components/student-table-shell.tsx
@@ -4,8 +4,8 @@ import { useRouter } from 'next/navigation'
 import { useCallback } from 'react'
 import { PaginationBar } from '@/app/app/_components/pagination-bar'
 import { useUpdateSearchParams } from '@/app/app/_hooks/use-update-search-params'
+import { SortableTableHead } from '@/components/sortable-table-head'
 import { Table, TableBody, TableHead, TableHeader, TableRow } from '@/components/ui/table'
-import { SortableTableHead } from '../_lib/sortable-head'
 import type { DashboardFilters, DashboardStudent } from '../types'
 import { STUDENTS_PAGE_SIZE } from '../types'
 import { StudentStatusFilter } from './student-status-filter'
