+++ apps/web/app/app/admin/dashboard/students/[id]/_components/session-history-table.tsx
@@ -3,8 +3,8 @@
 import { useCallback } from 'react'
 import { PaginationBar } from '@/app/app/_components/pagination-bar'
 import { useUpdateSearchParams } from '@/app/app/_hooks/use-update-search-params'
+import { SortableTableHead } from '@/components/sortable-table-head'
 import { Table, TableBody, TableHead, TableHeader, TableRow } from '@/components/ui/table'
-import { SortableTableHead } from '../../../_lib/sortable-head'
 import type { SessionSort, StudentSession, StudentSessionFilters } from '../../../types'
 import { SESSIONS_PAGE_SIZE } from '../../../types'
 import { ClickableSessionRow } from './clickable-session-row'
