+++ apps/web/app/app/admin/internal-exams/_utils/format-expiry.ts
@@ -1,0 +1,5 @@
+export function formatExpiry(iso: string): string {
+  const d = new Date(iso)
+  if (Number.isNaN(d.getTime())) return iso
+  return d.toLocaleString('en-GB', { dateStyle: 'medium', timeStyle: 'short' })
+}
