+++ apps/web/components/sortable-table-head.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { SortableTableHead } from './sortable-head'
+import { SortableTableHead } from './sortable-table-head'
 
 beforeEach(() => {
   vi.resetAllMocks()
